systemd/timer: remove timer persistency
This would cause the `nvchecker.service` unit to be started pretty much
instantly after the computer was turned off for more than an hour.

This is especially annoying for me, because my bridged network setup takes a
couple of seconds to be connected to the internet, so the nvchecker service
would just spill a bunch of connection errors and polute the system journal.