ci: run nvchecker using a pre-built container image
The container image is fairly light in size and already has all the
tools we need, so with that we avoid doing extra stuff, like compiling
nvchecker dependencies, and updating the distro repositories to install
other tools, like curl and xmlstarlet.

This lowers the average CI runtime in half (~35s to ~16s)!